% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_metric_files}
\alias{append_metric_files}
\title{Append combined metric data onto the data source}
\usage{
append_metric_files(current_combo_file, existing_combo_file, overwrite = FALSE)
}
\arguments{
\item{current_combo_file}{path to .csv file produced by \code{\link{merge_metric_files}}}

\item{existing_combo_file}{path to .csv file of all passed combo files}

\item{overwrite}{logical indicating if function should overwrite existing_combo_file with new combined file}
}
\value{
invisibly returns the combined data. Also it saves existing_combo_file with a datestamp
        for the archives and saves the output over existing_combo_file
}
\description{
Append combined metric data onto the data source
}
\examples{
\dontrun{
  #add examples to me please,
}
}
