% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{calc_num_tests}
\alias{calc_num_tests}
\title{INTERNAL function to calculate total tests per county per day}
\usage{
calc_num_tests(bcd, lab)
}
\arguments{
\item{bcd}{data.frame}

\item{lab}{data.frame}
}
\value{
data.frame
}
\description{
INTERNAL function to calculate total tests per county per day
}
\examples{
\dontrun{
  #write me an example
}
}
