% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{calc_pos_neg}
\alias{calc_pos_neg}
\title{INTERNAL function to calculate ingredients for percent positive per county per day}
\usage{
calc_pos_neg(lab)
}
\arguments{
\item{lab}{data.frame}
}
\value{
data.frame
}
\description{
INTERNAL function to calculate ingredients for percent positive per county per day
}
\examples{
\dontrun{
  #write me an example
}
}
