% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_nonneg}
\alias{check_nonneg}
\title{Checks that input values are non-negative, non-missing integers}
\usage{
check_nonneg(val, arg.name)
}
\arguments{
\item{val}{value}

\item{arg.name}{argument name for warning/error message.}
}
\value{
value
}
\description{
Checks that input values are non-negative, non-missing integers
}
