% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{clean_cli}
\alias{clean_cli}
\title{Clean ESSENCE data for CLI metrics}
\usage{
clean_cli(cli)
}
\arguments{
\item{cli}{data.frame from \code{\link{pull_essence}}}
}
\value{
a cleaned data.frame
}
\description{
Clean ESSENCE data for CLI metrics
}
\examples{
\dontrun{
  #write me an example please
}
}
