% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{clean_histTable}
\alias{clean_histTable}
\title{INTERNAL function to clean case data for metrics}
\usage{
clean_histTable(hdt, end_date)
}
\arguments{
\item{hdt}{a data.frame from the historical data table}

\item{end_date}{(default = NULL) If specified it is the end date of the
time series that you wish to analyze. It should be something
coercible to Date format with as.Date. If you want all data
leave it blank.}
}
\value{
raw historical case data ready for cleaning
}
\description{
INTERNAL function to clean case data for metrics
}
