% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{clean_hospital}
\alias{clean_hospital}
\title{Clean Hospital Data for metric calculation}
\usage{
clean_hospital(hosp, end_date)
}
\arguments{
\item{hosp}{a data.frame from EMResource}

\item{end_date}{(default = NULL) If specified it is the end date of the
time series that you wish to analyze. It should be something
coercible to Date format with as.Date. If you want all data
leave it blank.}
}
\value{
a cleaned data.frame
}
\description{
Clean Hospital Data for metric calculation
}
\examples{
\dontrun{
  #example here
}
}
