% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{clean_ili}
\alias{clean_ili}
\title{Clean ESSENCE data for ILI metrics}
\usage{
clean_ili(ili)
}
\arguments{
\item{ili}{data.frame from \code{\link{pull_essence}}}
}
\value{
a cleaned data.frame
}
\description{
Clean ESSENCE data for ILI metrics
}
\examples{
\dontrun{
  #write me an example please
}
}
