% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{clean_testing}
\alias{clean_testing}
\title{Internal function to clean testing data}
\usage{
clean_testing(lab, stg, end_date)
}
\arguments{
\item{lab}{data.frame from Lab table pull}

\item{stg}{data.frame from Staging table pull}

\item{end_date}{(default = NULL) If specified it is the end date of the
time series that you wish to analyze. It should be something
coercible to Date format with as.Date. If you want all data
leave it blank.}
}
\value{
a data.frame
}
\description{
Internal function to clean testing data
}
\examples{
\dontrun{
  #write me an example
}
}
