% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-be-deprecated.R
\name{clean_total_ed}
\alias{clean_total_ed}
\title{DEPRECATED - Clean ESSENCE data for Total ED metrics}
\usage{
clean_total_ed(total_ed)
}
\arguments{
\item{total_ed}{data.frame from \code{\link{pull_essence}}}
}
\value{
a cleaned data.frame
}
\description{
DEPRECATED - Clean ESSENCE data for Total ED metrics
}
\examples{
\dontrun{
  #write me an example please
}
}
