% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{essence_data}
\alias{essence_data}
\title{Call ESSENCE API and ingest response as data.frame}
\usage{
essence_data(url)
}
\arguments{
\item{url}{ESSENCE API url to pull either ILI or CLI data}
}
\value{
a data.frame from ESSENCE
}
\description{
Requires proper setup of ESSENCE credentials via the
\code{\link[keyring]{keyring}} package. See internal
DHS documentation about how to do that.
}
\examples{
\dontrun{
  #example here
}
}
