% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{essence_query}
\alias{essence_query}
\title{Customize ESSENCE API query}
\usage{
essence_query(url, start_date, end_date)
}
\arguments{
\item{url}{ESSENCE API url to pull either ILI or CLI data}

\item{start_date}{start date for daily time series}

\item{end_date}{end date for daily time series5}
}
\value{
a list of API calls with corrected dates
}
\description{
Customize ESSENCE API query
}
\examples{
\dontrun{
  #example here
}
}
