% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_dates}
\alias{fill_dates}
\title{Ensure that all units of observation have all dates within the max and min
of the date variable.}
\usage{
fill_dates(df, grouping_vars, date_var)
}
\arguments{
\item{df}{the data.frame to fill in}

\item{grouping_vars}{a character vector with the variable
names that define the grouping structure that will
uniquely identify the relavant units of observation}

\item{date_var}{}
}
\value{
the same data.frame with any required dates filled in for each
        unit. Variables without grouping_vars will be missing.
}
\description{
Ensure that all units of observation have all dates within the max and min
of the date variable.
}
\examples{
set.seed(123)
tmp <- tibble(County = sample(LETTERS[1:3], 15, replace = TRUE),
              Date = sample(seq(as.Date("2020-07-01"), as.Date("2020-07-07"), by = 1),
                            15, replace = TRUE)) \%>\%
  distinct(.) \%>\%
  arrange(County, Date)

#Compare

tmp

#To

arrange(fill_dates(tmp, grouping_vars = "County", date_var = "Date"),
        County, Date)

}
