% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hosp_sim}
\alias{hosp_sim}
\title{Fake hospitalization data inspired by the source data for hospitalization metrics}
\format{
A data.frame with ??? observations ??? variables:
\describe{
  \item{}{}
}
}
\usage{
hosp_sim
}
\description{
These data are fake
}
\keyword{datasets}
