% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_metric_files}
\alias{merge_metric_files}
\title{Combine metric data into a single output file}
\usage{
merge_metric_files(case, hosp, test, cli, ili, outfile)
}
\arguments{
\item{case}{data.frame produced by \code{\link{pull_histTable}} or \code{\link{pull_wedss}}}

\item{hosp}{data.frame produced by \code{\link{pull_hospital}}}

\item{test}{data.frame produced by}

\item{cli}{data.frame produced by}

\item{ili}{data.frame produced by}

\item{outfile}{file name (including path) for output data file}
}
\value{
invisibly returns the combined data
}
\description{
Combine metric data into a single output file
}
\examples{
\dontrun{
  #add examples to me please,
}
}
