% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-be-deprecated.R
\name{OLD_process_confirmed_cases}
\alias{OLD_process_confirmed_cases}
\title{OBSOLETE - VERSION Process the shaped confirmed case data.frame into a Tableau ready format}
\usage{
OLD_process_confirmed_cases(clean_case_df)
}
\arguments{
\item{clean_case_df}{shaped case data produced by \code{\link{shape_case_data}}}
}
\value{
a Tableau ready data.frame with the following columns:
\describe{
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
}
}
\description{
OBSOLETE - VERSION Process the shaped confirmed case data.frame into a Tableau ready format
}
\examples{
library(dplyr)

output <- pull_histTable() \%>\%
  shape_case_data() \%>\%
  process_confirmed_cases()
}
