% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{process_cli}
\alias{process_cli}
\title{Process the shaped CLI data into a Tableau ready format}
\usage{
process_cli(clean_cli_df)
}
\arguments{
\item{clean_cli_df}{### produced by \code{\link{shape_cli_data}}}
}
\value{
a Tableau ready data.frame with the following columns:
\describe{
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
}
}
\description{
Process the shaped CLI data into a Tableau ready format
}
\examples{
\dontrun{
  #write me an example
}
}
