% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{process_confirmed_cases}
\alias{process_confirmed_cases}
\title{Process the shaped confirmed case data.frame into a Tableau ready format}
\usage{
process_confirmed_cases(case_df)
}
\arguments{
\item{case_df}{Confirmed case data.frame (e.g. produced by \link{pull_histTable})}
}
\value{
a Tableau ready data.frame with the following columns:
\describe{
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
}
}
\description{
Process the shaped confirmed case data.frame into a Tableau ready format
}
\examples{
library(dplyr)

output <- pull_histTable() \%>\%
  shape_case_data() \%>\%
  process_confirmed_cases()
}
