% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{process_ili}
\alias{process_ili}
\title{Process the shaped ILI data into a Tableau ready format}
\usage{
process_ili(clean_ili_df, ili_threshold_path)
}
\arguments{
\item{clean_ili_df}{### produced by \code{\link{shape_ili_data}}}

\item{ili_threshold_path}{path to .csv file containing ILI thresholds}
}
\value{
a Tableau ready data.frame with the following columns:
\describe{
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
}
}
\description{
Process the shaped ILI data into a Tableau ready format
}
\examples{
\dontrun{
  #write me an example
}
}
