% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{process_testing}
\alias{process_testing}
\title{Process the shaped hospital data into a Tableau ready format}
\usage{
process_testing(clean_testing_df)
}
\arguments{
\item{clean_testing_df}{a list shaped case data produced by \code{\link{shape_testing_data}}}
}
\value{
a Tableau ready data.frame with the following columns:
\describe{
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
}
}
\description{
Process the shaped hospital data into a Tableau ready format
}
\examples{
\dontrun{
  #write me an example
}
}
