% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metrics.R
\name{process_total_ed}
\alias{process_total_ed}
\title{Process the shaped Total Emergency Department Visit data into a Tableau ready format}
\usage{
process_total_ed(total_ed_df)
}
\arguments{
\item{total_ed_df}{data.frame produced by \code{\link{pull_essence}} for
Total ED metrics}
}
\value{
a Tableau ready data.frame with the following columns:
\describe{
  \item{Region}{Name of geography}
  \item{Region_ID}{FIPS Code and/or region identifier}
  \item{Date}{Date of emergency dept visit}
  \item{RowType}{Are row values summary or daily values}
  \item{Total_ED_Visits}{Total ED visits for the day}
}
}
\description{
Process the shaped Total Emergency Department Visit data into a Tableau ready format
}
\examples{
\dontrun{
  #write me an example
}
}
