% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{pull_essence}
\alias{pull_essence}
\title{Pulls Essence Data for 3 metrics: CLI, ILI, Total ED Visits}
\usage{
pull_essence(
  api_url,
  start_date,
  end_date = NULL,
  metric = c("cli", "ili", "total_ed")
)
}
\arguments{
\item{api_url}{character string matching Essence API format}

\item{start_date}{Start date of the time series that you wish to analyze.
It should be something coercible to Date format with as.Date.}

\item{end_date}{(default = NULL) If specified it is the end date of the
time series that you wish to analyze. It should be something
coercible to Date format with as.Date. If you want all data
leave it blank.}

\item{metric}{one of "cli", "ili", "total_ed" depending on which metrics you
wish to calculate.}
}
\value{
a data.frame
}
\description{
Pulls Essence Data for 3 metrics: CLI, ILI, Total ED Visits
}
\examples{
\dontrun{
  #write me an example please
}
}
