% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{pull_histTable}
\alias{pull_histTable}
\title{Pull down DHS COVID-19 Historical Data Table}
\usage{
pull_histTable()
}
\value{
a cleaned version of the COVID-19 Historical Data Table including
HERC regions with the following columns
  \describe{
    \item{fips}{renamed from GEOID}
    \item{geo_type}{renamed from GEO}
    \item{geo_name}{renamed from NAME}
    \item{post_date}{LoadDttm converted to Date format}
    \item{case_daily}{cleaned daily new positive cases from POS_NEW except for the first day which is from POSITIVE}
    \item{test_daily}{cleaned daily new total tests from TEST_NEW except for the first day which is from POSITIve + NEGATIVE}
    \item{death_daily}{cleaned daily new deaths from DTH_NEW except for the first day which is from DEATHS}
    \item{pop_2018}{2018 Population Numbers pulled from WISH}
    \item{case_cum}{daily cumulative positive cases calculated from case_daily}
    \item{test_cum}{daily cumulative total tests calculated from test_daily}
    \item{death_cum}{daily cumulative deaths calculated from death_daily}
  }
  and likely one or more of the following columns if applicable
  \describe{
    \item{case_daily_raw}{original daily new positive cases before cleaning}
    \item{test_daily_raw}{original daily new total tests before cleaning}
    \item{death_daily_raw}{cleaned daily new deaths before cleaning}
  }
}
\description{
This function uses the REST API to pull down the data that are currently
posted at https://data.dhsgis.wi.gov/datasets/covid-19-historical-data-table/
using the GeoJSON REST API. \emph{Note: currently it does not pull down any
geometry data used to produce maps.}
}
\examples{
hdt <- pull_histTable()
}
