% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{pull_hospital}
\alias{pull_hospital}
\title{Imports Hospitalization file produced from EM Resource}
\usage{
pull_hospital(file, end_date)
}
\arguments{
\item{file}{path to EM Resource hospitalization summary .csv file}

\item{end_date}{(default = NULL) If specified it is the end date of the
time series that you wish to analyze. It should be something
coercible to Date format with as.Date. If you want all data
leave it blank.}
}
\value{
a data.frame suitable for metric calculations
}
\description{
Imports Hospitalization file produced from EM Resource
}
\examples{
\dontrun{
  pull_hospital("hospdtatafile.csv")
}
}
