% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{pull_wedss}
\alias{pull_wedss}
\title{Pull Confirmed Case data from WEDSS}
\usage{
pull_wedss(query, conn, end_date)
}
\arguments{
\item{query}{query to pull data from WEDSS. This call should only pull
counties or jurisdictions that will automatically be
aggregated into HERC regions and Statewide.}

\item{conn}{connection to database (see \code{\link[RODBC]{odbcConnect}})}

\item{end_date}{(default = NULL) If specified it is the end date of the
time series that you wish to analyze. It should be something
coercible to Date format with as.Date. If you want all data
leave it blank.}
}
\value{
a cleaned version of the COVID-19 Historical Data Table including
HERC regions with the following columns
  \describe{
    \item{fips}{renamed from GEOID}
    \item{geo_type}{renamed from GEO}
    \item{geo_name}{renamed from NAME}
    \item{post_date}{LoadDttm converted to Date format}
    \item{case_daily}{cleaned daily new positive cases from POS_NEW except for the first day which is from POSITIVE}
    \item{test_daily}{cleaned daily new total tests from TEST_NEW except for the first day which is from POSITIve + NEGATIVE}
    \item{death_daily}{cleaned daily new deaths from DTH_NEW except for the first day which is from DEATHS}
    \item{pop_2018}{2018 Population Numbers pulled from WISH}
    \item{case_cum}{daily cumulative positive cases calculated from case_daily}
    \item{test_cum}{daily cumulative total tests calculated from test_daily}
    \item{death_cum}{daily cumulative deaths calculated from death_daily}
  }
  and likely one or more of the following columns if applicable
  \describe{
    \item{case_daily_raw}{original daily new positive cases before cleaning}
    \item{test_daily_raw}{original daily new total tests before cleaning}
    \item{death_daily_raw}{cleaned daily new deaths before cleaning}
  }
}
\description{
This function supplies a wrapper to fetch the results from a SQL query
from WEDSS and then performs basic data cleaning to calculate the
confirmed case metrics.
}
\examples{
\dontrun{
   library(RODBC)
   channel <- odbcConnect("databasename", ...)

   sql_query <- "SELECT * FROM TABLE WHERE GEO = 'COUNTY'"

   hdt <- pull_wedss(query = sql_query, conn = channel, end_date = as.Date("2020-06-24"))
}

}
