% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{pval_trajectory}
\alias{pval_trajectory}
\title{Calculate p-value of trajectory current counts vs prior counts}
\usage{
pval_trajectory(curr, prev)
}
\arguments{
\item{curr}{vector of counts for current period (must be non-negative integer)}

\item{prev}{vector of counts for previous period (must be non-negative integer)}
}
\value{
a p.value extracted from the results of \code{\link[stats]{poisson.test}}
}
\description{
Calculate p-value of trajectory current counts vs prior counts
}
\examples{
traj_pval <- pval_trajectory(curr = 100L, prev = 80L)
}
