% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{rev_cusum_lcl}
\alias{rev_cusum_lcl}
\title{Calculates lower control limits for the reverse cusum control chart based on a
3\eqn{\sigma} significance level.}
\usage{
rev_cusum_lcl(curr, delta_t = 1)
}
\arguments{
\item{curr}{vector of counts for current period (must be non-negative integer)}

\item{delta_t}{number of time periods back in time.}
}
\value{
the greatest count smaller than curr that would produce a statistically
        significant \code{\link[stats]{poisson.test}}
}
\description{
Calculates lower control limits for the reverse cusum control chart based on a
3\eqn{\sigma} significance level.
}
\examples{
#For 3 weeks into the past using 7-day binned counts
rev_cusum_lcl(curr = 100L, delta_t = 3)
}
