% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{score_burden}
\alias{score_burden}
\title{calculate the burden}
\usage{
score_burden(curr, prev, pop)
}
\arguments{
\item{curr}{vector of counts for current period (must be non-negative integer)}

\item{prev}{vector of counts for previous period (must be non-negative integer)}

\item{pop}{vector of population values to standardize counts}
}
\value{
a numeric vector of population adjusted burden
}
\description{
calculate the burden
}
\examples{
burd <- score_burden(curr = 100L, prev = 80L, population = 65432)
}
