% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{score_trajectory}
\alias{score_trajectory}
\title{Trajectory Measure for Count Data}
\usage{
score_trajectory(curr, prev)
}
\arguments{
\item{curr}{count for current period (must be non-negative integer)}

\item{prev}{count for previous period (must be non-negative integer)}
}
\value{
a numeric scalar
}
\description{
Calculates the ratio of the most current to previous period
counts. Values greater than 1 indicate growing; less than one
indicate shrinking. \emph{NOTE: Counts must have the same period
(e.g. both summed over 7 consecutive days.)} In order to avoid
extreme values and dividing by zeros, the ratio includes unit offset
for pseudocounts and is capped at 5 (i.e. a 400\% increase).
}
\examples{
traj_2020_06_10 <- score_trajectory(curr = 100L, prev = 80L)
}
