% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{shape_case_data}
\alias{shape_case_data}
\title{Reshape confirmed case data for producing Tableau extracts}
\usage{
shape_case_data(case_df)
}
\arguments{
\item{case_df}{Confirmed case data.frame (e.g. produced by \link{pull_histTable})}
}
\value{
a data.frame with the following with one row per county, state, and
HERC regions with the following columns
\describe{
  \item{fips}{FIPS Code and/or region identifier}
  \item{geo_name}{Name of geography}
  \item{pop_2018}{2018 Population Numbers pulled from WISH}
  \item{case_weekly_1}{Total cases for \strong{current} 7 day period}
  \item{case_weekly_2}{Total cases for \strong{prior} 7 day period}
  \item{week_end_1}{End date for \strong{current} 7 day period}
  \item{week_end_2}{End date for \strong{prior} 7 day period}
}
}
\description{
take case data (from WEDSS or historical data table) and put it in proper
shape for metric tables to feed tableau
}
\examples{
\dontrun{
  hdt <- pull_histTable()
  hdt_clean <- shape_case_data(hdt)
}
}
