% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metrics.R
\name{shape_cli_data}
\alias{shape_cli_data}
\title{Shape CLI data}
\usage{
shape_cli_data(cli_df)
}
\arguments{
\item{cli_df}{data.frame produced by \code{\link{pull_essence}} for
CLI metrics}
}
\value{
a list of data.frames. The "summary" data.frame has one row per
county, state, and HERC region with the following columns
\describe{
  \item{Visit_Date}{Date of emergency dept visit}
  \item{fips}{FIPS Code and/or region identifier}
  \item{County}{Name of geography}
  \item{RowType}{Are row values summary or daily values}
  \item{DailyED}{Count of CLI ED visits for the day}
  \item{}{}
  \item{}{}
  \item{}{}
}
and the "daily" data.frame has one row per county, state, and HERC region
per day for the two week period with the following columns
\describe{
  \item{fips}{FIPS Code and/or region identifier}
  \item{County}{Name of geography}
  \item{pop_2018}{2018 Population Numbers pulled from WISH}
  \item{WeeklyED_1}{Total CLI visits for \strong{current} 7 day period}
  \item{WeeklyED_2}{Total CLI visits for \strong{previous} 7 day period}
  \item{week_end_1}{End date for \strong{current} 7 day period}
  \item{week_end_2}{End date for \strong{prior} 7 day period}
  \item{RowType}{Are row values summary or daily values}
}
}
\description{
Shape CLI data
}
\examples{
\dontrun{
  #write me an example
}
}
