% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metrics.R
\name{shape_hospital_data}
\alias{shape_hospital_data}
\title{Shape EM Resource summary data for metric calculations}
\usage{
shape_hospital_data(hosp_df)
}
\arguments{
\item{hosp_df}{data.frame output by \code{\link{pull_hospital}}}
}
\value{
a list of data.frames. The "summary" data.frame has one row per
county, state, and HERC regions with the following columns
\describe{
  \item{Run_Date}{Date hospital report was run}
  \item{RowType}{Are row values summary or daily values}
  \item{fips}{FIPS Code and/or region identifier}
  \item{geo_name}{Name of geography}
  \item{pop_2018}{2018 Population Numbers pulled from WISH}
  \item{covid_reg_weekly_1}{Hospitalized Covid cases for \strong{current} 7 day period}
  \item{covid_reg_weekly_2}{Hospitalized Covid cases for \strong{prior} 7 day period}
  \item{covid_icu_weekly_1}{ICU Covid cases for \strong{current} 7 day period}
  \item{covid_icu_weekly_2}{ICU Covid cases for \strong{prior} 7 day period}
  \item{week_end_1}{End date for \strong{current} 7 day period}
  \item{week_end_2}{End date for \strong{prior} 7 day period}
}
and the "daily" data.frame has one row per state and HERC region (\emph{not county})
per day for the two week period with the following columns
\describe{
  \item{County}{Name of geography}
  \item{Report_Date}{Date hospital information was reported}
  \item{dailyCOVID_px}{}
  \item{dailyCOVID_ICUpx}{}
  \item{geo_type}{Type of geography - all missing should be removed}
  \item{fips}{FIPS Code and/or region identifier}
  \item{pop_2018}{WISH 2018 population figures}
  \item{RowType}{Are row values summary or daily values}
  \item{totalbeds}{Total Beds (ICU, Intermediate, Med/Surg, Neg. Flow)}
  \item{beds_IBA}{Total Immediate Beds Available (ICU, Intermediate, Med/Surg, Neg. Flow)}
  \item{totalICU}{Total ICU Beds}
  \item{ICU_IBA}{Immediate ICU Beds Available}
  \item{num_px_vent}{Current Number of Ventilated Patients (Induvated and mechanically ventilated)}
  \item{total_vents}{Total rented/owned/demoed general use ventilators on hand}
  \item{intermed_beds_IBA}{Immediate Intermediate Care Beds Available}
  \item{negflow_beds_IBA}{Immediate Negative Airflow Isolation Beds Available}
  \item{medsurg_beds_IBA}{Immediate Medical/Surgical Beds Available}
  \item{PrctBeds_IBA}{Hosp_beds_IBA / Hosp_totalbeds}
  \item{PrctICU_IBA}{Hosp_ICU_IBA / Hosp_ICU_IBA}
  \item{PrctVent_Used}{Hosp_num_px_vent / Hosp_total_vents}
  \item{Run_Date}{Date hospital report was run}
}
}
\description{
Shape EM Resource summary data for metric calculations
}
\examples{
\dontrun{
  #Add examples here
}
}
