% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{shape_ili_data}
\alias{shape_ili_data}
\title{Shape ILI data}
\usage{
shape_ili_data(ili_df)
}
\arguments{
\item{ili_df}{data.frame produced by \code{\link{pull_essence}} for
ILI metrics}
}
\value{
output
}
\description{
Shape ILI data
}
\examples{
\dontrun{
  #write me an example
}
}
