% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metrics.R
\name{shape_testing_data}
\alias{shape_testing_data}
\title{Shape Testing summary data for metric calculations}
\usage{
shape_testing_data(testing_df)
}
\arguments{
\item{testing_df}{data.frame output by \code{\link{pull_testing}}}
}
\value{
a list of data.frames. The "summary" data.frame has one row per
county, state, and HERC region with the following columns
\describe{
  \item{Region_ID}{FIPS Code and/or region identifier}
  \item{Region}{Name of geography}
  \item{RowType}{Are row values summary or daily values}
  \item{Positives}{Count of positive specimens}
  \item{Negatives}{Count of negative specimens}
  \item{Total}{Positives + Negatives}
  \item{Date}{Summary Date for 2 week period}
}
and the "daily" data.frame has one row per county, state, and HERC region
per day for the two week period with the following columns
\describe{
  \item{Date}{Date test results came in}
  \item{Region_ID}{FIPS Code and/or region identifier}
  \item{Region}{Name of geography}
  \item{RowType}{Are row values summary or daily values}
  \item{Positives}{Count of positive specimens}
  \item{Negatives}{Count of negative specimens}
  \item{Total}{Positives + Negatives}
}
}
\description{
Shape Testing summary data for metric calculations
}
\examples{
\dontrun{
  #Add examples here
}
}
