% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std_region}
\alias{std_region}
\title{Utility to standardize geographic names}
\usage{
std_region(reg)
}
\arguments{
\item{reg}{vector of geographic names}
}
\value{
a vector of standardized geographic names
}
\description{
Utility to standardize geographic names
}
