% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std_region_id}
\alias{std_region_id}
\title{Utility to standardize geographic id's}
\usage{
std_region_id(reg_id)
}
\arguments{
\item{reg_id}{vector of geographic ids}
}
\value{
a vector of standardized geographic ids
}
\description{
Utility to standardize geographic id's
}
