% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{calc_num_tests}
\alias{calc_num_tests}
\title{INTERNAL function to calculate total tests per county per day}
\usage{
calc_num_tests(bcd, lab, end_date)
}
\arguments{
\item{bcd}{data.frame}

\item{lab}{data.frame}

\item{end_date}{(default = NULL) If specified it is the end date of the
time series that you wish to analyze. It should be something
coercible to Date format with as.Date. If you want all data
leave it blank.}
}
\value{
data.frame
}
\description{
INTERNAL function to calculate total tests per county per day
}
\examples{
\dontrun{
  #write me an example
}
}
