% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{class_burden}
\alias{class_burden}
\title{Classify population adjusted burden}
\usage{
class_burden(burden)
}
\arguments{
\item{burden}{numeric vector as calculated by \code{\link{score_burden}}}
}
\value{
an ordered factor
}
\description{
Currently, cutoff values are hard coded and adopted straight from
the ones used in the CDC State Indicators report:
\itemize{
  \item "Low" is a burden less than or equal to 10 per 100,000
  \item "Moderate" is a burden greater than 10 and less than or equal to 50 per 100,000
  \item "Moderately high" is a burden greater than 50 and less than or equal to 100 per 100,000
  \item "High" is a burden greater than 100 and less than or equal to 350 per 100,000
  \item "Very high" is a burden greater than 350 per 100,000
}
}
\examples{
burd <- score_burden(curr = 100L, prev = 80L, pop = 65432L)
burd_class <- class_burden(burd)
}
