% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{class_trajectory}
\alias{class_trajectory}
\title{Classify Trajectory Measure for County Data}
\usage{
class_trajectory(traj, pval)
}
\arguments{
\item{traj}{trajectory as calculated by \code{\link{score_trajectory}}}

\item{pval}{p-value as calculated by \code{\link{pval_trajectory}}}
}
\value{
an ordered factor "Shrinking" < "Not Statistically Significant" < "Growing"
}
\description{
This function tests to see if the count trajectory is statistically
distinct from zero with a Poisson test for the ratio of two counts
using \code{\link[stats]{poisson.test}} as.calculated by \code{\link{pval_trajectory}}
at the two-sided p < .05 level. If it is and the ratio as calculated by
\link{score_trajectory} is less than or equal to 0.9, the trajectory is
classified as shrinking. If it is and the ration is greather than or equal
to 1.1, the trajectory is classified as growing. If the test fails to
reject the null hypothesis, the trajectory is classified as not statistically
significant.
}
\examples{
traj_score <- score_trajectory(curr = 100L, prev = 80L)
traj_pval  <- pval_trajectory(curr = 100L, prev = 80L)

traj_class_2020_06_10 <- class_trajectory(traj_score, traj_pval)
}
