% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{confirmed_case_composite}
\alias{confirmed_case_composite}
\title{Calculate the confirmed case composite summary}
\usage{
confirmed_case_composite(traj_class, burd_class)
}
\arguments{
\item{traj_class}{trajectory classification as calculated by \code{\link{class_trajectory}}}

\item{burd_class}{burden classification as calculated by \code{\link{class_burden}}}
}
\value{
an ordered factor for the composite confirmed case summary
}
\description{
Source: Jeff's Design Doc as confirmed by SEOC
}
\examples{
burd <- score_burden(curr = 100L, prev = 80L, population = 65432)
burd_c <- class_burden(burd)
traj_score <- score_trajectory(curr = 100L, prev = 80L)
traj_pval  <- pval_trajectory(curr = 100L, prev = 80L)
traj_class_2020_06_10 <- class_trajectory(traj_score, traj_pval)

comp_2020_06_10 <- confirmed_case_composite(traj_class_2020_06_10, burd_c)
}
