% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{fdr_trajectory}
\alias{fdr_trajectory}
\title{Calculates the False Discovery Rate for trajectory p-values}
\usage{
fdr_trajectory(pval)
}
\arguments{
\item{pval}{a vector of p-values}
}
\value{
the false discovery rate
}
\description{
According to method of Benjamini & Hochberg (1995) via \code{\link[stats]{p.adjust}}
}
\examples{
pval_2020_06_10 <- pval_trajectory(curr = 100L, prev = 80L)
fdr_2020_06_10 <- fdr_trajectory(pval_2020_06_10)
}
