% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{process_hospital}
\alias{process_hospital}
\title{Process the shaped hospital data into a Tableau ready format}
\usage{
process_hospital(clean_hosp_df)
}
\arguments{
\item{clean_hosp_df}{a list of shaped hospital daily and summary
data produced by \code{\link{shape_hospital_data}}}
}
\value{
a Tableau ready data.frame with the following columns:
\describe{
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
}
}
\description{
Process the shaped hospital data into a Tableau ready format
}
\examples{
\dontrun{
library(dplyr)

output <- pull_hospital("path-to-hospital-extract") \%>\%
  shape_hospital_data() \%>\%
  process_hospital()
}
}
