% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metrics.R
\name{process_ili}
\alias{process_ili}
\title{Process the shaped ILI data into a Tableau ready format}
\usage{
process_ili(ili_df, ili_threshold_path)
}
\arguments{
\item{ili_df}{data.frame produced by \code{\link{pull_essence}} for
ILI metrics}

\item{ili_threshold_path}{path to .csv file containing ILI thresholds}
}
\value{
a Tableau ready data.frame with the following columns:
\describe{
  \item{Region}{Name of geography}
  \item{Region_ID}{FIPS Code and/or region identifier}
  \item{Date}{Date of emergency dept visit}
  \item{RowType}{Are row values summary or daily values}
  \item{ED_Total_Visits}{Total visits to ED}
  \item{ED_ILI_Visits}{Count of ILI ED visits}
  \item{ILI_Percent}{Percent of ED visits related to ILI}
  \item{ILI_Baseline}{Baseline value for ILI}
  \item{ILI_Threshold}{Threshold value for ILI}
  \item{ILI_Status}{Summary ILI metric}
  \item{ILI_Moving_Avg}{right-aligned 7 day rolling average of ILI_Percent}
}
}
\description{
Process the shaped ILI data into a Tableau ready format
}
\examples{
\dontrun{
  #write me an example
}
}
