% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metrics.R
\name{process_testing}
\alias{process_testing}
\title{Process the shaped testing data into a Tableau ready format}
\usage{
process_testing(testing_df)
}
\arguments{
\item{testing_df}{data.frame output by \code{\link{pull_testing}}}
}
\value{
a Tableau ready data.frame with the following columns:
\describe{
  \item{Date}{Date hospital information was reported}
  \item{Region}{Name of geographic unit (county, state, HERC region)}
  \item{Region_ID}{ID code for geographic unit (FIPS for county and state)}
  \item{RowType}{Are row values summary or daily values}
  \item{Testing_Positive_Specimens}{Count of specimens positive for COVID-19}
  \item{Testing_Nonpositive_Specimens}{Count of specimens not positive for COVID-19}
  \item{Testing_Total_Specimens}{Count of specimens}
  \item{Testing_Percent_Positive}{Percent specimens positive for COVID-19}
  \item{Testing_Incident_Tests}{Total number of incidents tested}
  \item{Testing_Incident_Test_Target}{Target number of incidents to test}
  \item{Testing_Percent_of_Target}{Percent incidents tested out of target (Testing_Incident_Tests / Testing_Incident_Test_Target)}
  \item{Testing_Composite_Class}{Summary metric for testing based on Testing_Percent_Positive and Testing_Percent_of_Target}
}
}
\description{
Process the shaped testing data into a Tableau ready format
}
\examples{
\dontrun{
  #write me an example
}
}
