% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{process_total_ed}
\alias{process_total_ed}
\title{Process the shaped Total Emergency Department Visit data into a Tableau ready format}
\usage{
process_total_ed(total_ed_df)
}
\arguments{
\item{total_ed_df}{data.frame produced by \code{\link{pull_essence}} for
Total ED metrics}
}
\value{
a Tableau ready data.frame with the following columns:
\describe{
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
}
}
\description{
Process the shaped Total Emergency Department Visit data into a Tableau ready format
}
\examples{
\dontrun{
  #write me an example
}
}
