% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{shape_cli_data}
\alias{shape_cli_data}
\title{Shape CLI data}
\usage{
shape_cli_data(cli_df)
}
\arguments{
\item{cli_df}{data.frame produced by \code{\link{pull_essence}} for
CLI metrics}
}
\value{
output
}
\description{
Shape CLI data
}
\examples{
\dontrun{
  #write me an example
}
}
