% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{shape_hospital_data}
\alias{shape_hospital_data}
\title{Shape EM Resource summary data for metric calculations}
\usage{
shape_hospital_data(hosp_df)
}
\arguments{
\item{hosp_df}{data.frame output by \code{\link{pull_hospital}}}
}
\value{
a list of data.frames (one summary and one daily)
}
\description{
Shape EM Resource summary data for metric calculations
}
\examples{
\dontrun{
  #Add examples here
}
}
