% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metrics.R
\name{shape_ili_data}
\alias{shape_ili_data}
\title{Shape ILI data}
\usage{
shape_ili_data(ili_df)
}
\arguments{
\item{ili_df}{data.frame produced by \code{\link{pull_essence}} for
ILI metrics}
}
\value{
a list of data.frames. The "summary" data.frame has one row per
county, state, and HERC region with the following columns
\describe{
  \item{Region}{Name of geography}
  \item{Region_ID}{FIPS Code and/or region identifier}
  \item{Date}{Date of emergency dept visit}
  \item{RowType}{Are row values summary or daily values}
  \item{Total_Visits}{Total ED visits for the day}
  \item{ILI_Visits}{Count of ILI ED visits for the day}
}
and the "daily" data.frame has one row per county, state, and HERC region
per day for the two week period with the following columns
\describe{
  \item{Region}{Name of geography}
  \item{Region_ID}{FIPS Code and/or region identifier}
  \item{Date}{Date of emergency dept visit}
  \item{RowType}{Are row values summary or daily values}
  \item{Total_Visits}{Total ED visits for the day}
  \item{ILI_Visits}{Count of ILI ED visits for the day}
  \item{ILI_perc}{Percent of ED visits related to ILI}
}
}
\description{
Shape ILI data
}
\examples{
\dontrun{
  #write me an example
}
}
