% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metrics.R
\name{shape_testing_data}
\alias{shape_testing_data}
\title{Shape Testing summary data for metric calculations}
\usage{
shape_testing_data(testing_df)
}
\arguments{
\item{testing_df}{data.frame output by \code{\link{pull_testing}}}
}
\value{
a list of data.frames. The "summary" data.frame has one row per
county, state, and HERC region with the following columns
\describe{
  \item{RowType}{Are row values summary or daily values}
  \item{Region_ID}{FIPS Code and/or region identifier}
  \item{Area}{Name of geography}
  \item{Testing_Volume}{2 week testing volume target}
  \item{Tests}{Count of Incident Tests}
  \item{NotPositive}{Count of non-positive specimens}
  \item{Positive}{Count of positive specimens}
  \item{resultdateonly}{Date test results came in}
}
and the "daily" data.frame has one row per county, state, and HERC region
per day for the two week period with the following columns
\describe{
  \item{resultdateonly}{Date test results came in}
  \item{Area}{Name of geography}
  \item{Tests}{Count of Incident Tests}
  \item{NotPositive}{Count of non-positive specimens}
  \item{Positive}{Count of positive specimens}
  \item{Region_ID}{FIPS Code and/or region identifier}
  \item{RowType}{Are row values summary or daily values}
}
}
\description{
Shape Testing summary data for metric calculations
}
\examples{
\dontrun{
  #Add examples here
}
}
