% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_fcns.R
\name{shape_testing_data}
\alias{shape_testing_data}
\title{Shape Testing summary data for metric calculations}
\usage{
shape_testing_data(testing_df)
}
\arguments{
\item{testing_df}{data.frame output by \code{\link{pull_testing}}}
}
\value{
a list of data.frames (one summary and one daily)
}
\description{
Shape Testing summary data for metric calculations
}
\examples{
\dontrun{
  #Add examples here
}
}
