% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_metrics.R
\name{shape_total_ed_data}
\alias{shape_total_ed_data}
\title{Shape Total Emergency Department Visits data}
\usage{
shape_total_ed_data(total_ed_df)
}
\arguments{
\item{total_ed_df}{data.frame produced by \code{\link{pull_essence}} for
Total ED metrics}
}
\value{
output
}
\description{
Shape Total Emergency Department Visits data
}
\examples{
\dontrun{
  #write me an example
}
}
